refactor: split to swap context
